pyproject.toml
[tool.poetry]
name = "vucse"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "~4.0"
django-autoslug = "^1.9.9"
djangorestframework = "^3.14.0"
dj-database-url = "^2.1.0"
whitenoise = "^6.6.0"
django-jazzmin = "^2.6.0"
python-slugify = {extras = ["unidecode"], version = "^8.0.2"}
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
